[Refactor] UIManager에서 TitleUIManager 파일 분리
