+++ Assets/Script/UIManager.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bc7e9eff7f3a78e45bfea3d44cd3366d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
