+++ Assets/Script/UIManager/LobbyUIManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 08354109d807a8347bfa1e5c5fd1a2bc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
