+++ Assets/Script/UIManager/TitleUIManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c0bd9630571d5ef49be43ef895e66aa7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
